ARGOS satellite tracking data for an individual Pectoral Sandpiper

A dataset containing ARGOS locations for an individual male Pectoral Sandpiper

Usage

pesa56511

Format

A data.table with 882 rows and 4 columns:

latitude

Numeric. Latitude in decimal degrees (WGS84).

longitude

Numeric. Longitude in decimal degrees (WGS84).

locationDate

POSIXct. Timestamp of location fix (UTC).

locationClass

Character. ARGOS location quality class (e.g., B, 0, A).

Source

Kempenaers, Bart, and Mihai Valcu. "Breeding site sampling across the Arctic by individual males of a polygynous shorebird." Nature 541.7638 (2017): 528-531.

Examples

data(pesa56511)
head(pesa56511)
#>    latitude longitude        locationDate locationClass
#>       <num>     <num>              <POSc>         <ord>
#> 1:   71.522  -155.544 2014-06-02 21:46:07             B
#> 2:   70.860  -154.020 2014-06-02 22:14:52             0
#> 3:   71.314  -156.703 2014-06-02 22:25:07             0
#> 4:   71.336  -156.746 2014-06-02 22:46:43             0
#> 5:   71.325  -156.739 2014-06-02 23:25:19             B
#> 6:   71.330  -156.675 2014-06-02 23:51:33             2