ARGOS satellite tracking data for an individual Ruff

A dataset containing ARGOS locations for an individual male Ruff.

Usage

ruff143789

Format

For ruff143789: a data.table with 2834 rows and 4 columns.

For mini_ruff: a data.table with 276 rows and the same columns.

Source

Kempenaers, Bart, et al. "Large-scale sampling of potential breeding sites in male ruffs." Proceedings B 292.2038 (2025): 20242225. https://royalsocietypublishing.org/doi/full/10.1098/rspb.2024.2225.

Details

ruff143789 is the full dataset; mini_ruff is a reduced subset used for examples and tests.

Examples

data(ruff143789)
data(mini_ruff)
head(mini_ruff)
#>    latitude longitude                time
#>       <num>     <num>              <POSc>
#> 1:   66.751    41.098 2015-05-31 14:33:09
#> 2:   66.741    40.881 2015-05-31 15:12:44
#> 3:   66.670    40.801 2015-05-31 16:41:01
#> 4:   66.665    40.985 2015-05-31 16:50:30
#> 5:   66.427    40.462 2015-05-31 17:43:18
#> 6:   66.485    40.456 2015-05-31 18:20:13